backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rishav Kundu <rk@rishav.io>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.1.0"}
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
